mv getSecurityHeaders into security.ts.
